Normalize a Spectra2D Object

This function carries out normalization of the spectra in a Spectra2D object. The current options are:

"zero2one" normalizes each 2D spectrum to a [0 ... 1] scale.

"minusPlus" normalizes each 2D spectrum to a [-1 ... 1] scale.

"TotInt" normalizes each 2D spectrum so that the total area is one.

normSpectra2D(spectra, method = "zero2one")

Arguments

spectra

An object of S3 class Spectra2D to be normalized.

method

Character string giving the method for normalization.

Value

An object of S3 class Spectra2D.

See also

Author

Bryan A. Hanson, DePauw University.

Examples

data(MUD1)
MUD1n <- normSpectra2D(MUD1)
MUD1b <- removeFreq(MUD1, remF2 = 2.5 ~ 3.5)
MUD1bn <- normSpectra2D(MUD1b)