Spectra2D Objects

In ChemoSpec2D, spectral data sets are stored in an S3 class called Spectra2D, which contains a variety of information in addition to the spectra themselves. Spectra2D objects are created by files2Spectra2DObject.

Structure

The structure of a Spectra2D object is a list of eight elements and an attribute as follows: w

element

type

description

$F2	num	A common frequency (or wavelength) axis corresponding

to the F2 dimension in NMR or the x axis more generally.

Must be sorted ascending.

$F1	num

A common frequency (or wavelength) axis corresponding

to the F1 dimension in NMR or the y axis more generally.

Must be sorted ascending.

$data

num	A list of matrices. Each matrix contains a 2D spectrum.

Each matrix should have

length(F1)	rows and

length(F2)	columns. The low end of the F2

dimension is last column of the last row (lower right hand corner

as typically displayed). The low end of the F1 dimension

is the last column of the first row (upper right hand corner).

In other words, the spectrum is stored as typically displayed.

The list of matrices, if named, should have the same names as

names

. However, this is not currently enforced.

$names	chr	The sample names for the spectra; length must be no. samples.

$groups

factor	The group classification of the samples; length must be no. samples.

$colors	character

Colors for plotting; length must be no. samples.

Colors correspond to groups.

$units	chr

Three entries, the first giving the F2 (x) axis unit, the

second the F1 (y) axis unit, and the third the z axis unit,

usually some kind of intensity.

$desc

chr	A character string describing the data set.

- attr	chr

"Spectra2D" The S3 class designation.

See also

Examples

data(MUD1)
str(MUD1)

#> List of 8
#>  $ F2    : num [1:51] 0 0.2 0.4 0.6 0.8 1 1.2 1.4 1.6 1.8 ...
#>  $ F1    : num [1:41] 10 10.5 11 11.5 12 12.5 13 13.5 14 14.5 ...
#>  $ data  :List of 10
#>   ..$ Sample1 : num [1:41, 1:51] 2.35e-02 9.76e-03 -1.31e-02 2.07e-02 6.15e-05 ...
#>   ..$ Sample2 : num [1:41, 1:51] 0.0275 -0.0377 -0.0548 0.0968 -0.0141 ...
#>   ..$ Sample3 : num [1:41, 1:51] -0.03273 0.00819 0.00402 -0.0162 0.06004 ...
#>   ..$ Sample4 : num [1:41, 1:51] -0.0784 -0.0368 0.0211 -0.01 -0.0217 ...
#>   ..$ Sample5 : num [1:41, 1:51] 0.09964 -0.00588 0.04498 -0.05829 -0.0356 ...
#>   ..$ Sample6 : num [1:41, 1:51] -0.0444 -0.0998 0.0477 0.0909 -0.041 ...
#>   ..$ Sample7 : num [1:41, 1:51] -0.0322 -0.0243 -0.0573 0.0539 0.0537 ...
#>   ..$ Sample8 : num [1:41, 1:51] 0.00781 -0.00185 0.01103 0.09974 -0.03219 ...
#>   ..$ Sample9 : num [1:41, 1:51] 0.086 -0.052 -0.0493 -0.0161 0.0698 ...
#>   ..$ Sample10: num [1:41, 1:51] -0.067148 0.000105 0.082326 0.099195 -0.074405 ...
#>  $ names : chr [1:10] "Sample1" "Sample2" "Sample3" "Sample4" ...
#>  $ groups: Factor w/ 2 levels "GroupA","GroupB": 1 1 1 1 1 2 2 2 2 2
#>  $ colors: chr [1:10] "red" "red" "red" "red" ...
#>  $ units : chr [1:3] "ppm" "ppm" "intensity"
#>  $ desc  : chr "MUD1: HSQC-like data for ChemoSpec2D demo & testing"
#>  - attr(*, "class")= chr "Spectra2D"

sumSpectra(MUD1)

#> 
#>  MUD1: HSQC-like data for ChemoSpec2D demo & testing 
#> 
#> 	There are 10 spectra in this set.
#> 
#> 	The F2 dimension runs from 0 to 10 ppm
#> 	and there are 51 data points.
#> 
#> 	The F1 dimension runs from 10 to 30 ppm
#> 	and there are 41 slices.
#> 
#> 	The spectra are divided into 2 groups: 
#> 
#>    group no. color
#> 1 GroupA   5   red
#> 2 GroupB   5  blue
#> 
#> 
#> *** Note: this is an S3 object
#> of class 'Spectra2D'