Import Data into a Spectra2D Object

This function imports data into a Spectra2D object. It uses read.table to read files so it is very flexible in regard to file formatting. Be sure to see the … argument below for important details you need to provide.

files2Spectra2DObject(gr.crit = NULL, gr.cols = "auto", fmt = NULL,
  nF2 = NULL, x.unit = "no frequency unit provided",
  y.unit = "no frequency unit provided",
  z.unit = "no intensity unit provided",
  descrip = "no description provided", fileExt = "\\.(csv|CSV)$",
  out.file = "mydata", debug = FALSE, ...)

Arguments

gr.crit
Group Criteria. A vector of character strings which will be searched for among the file/sample names in order to assign an individual spectrum to group membership. This is done using grep, so characters like "." (period/dot) do not have their literal meaning (see below). Warnings are issued if there are file/sample names that don't match entries in gr.crit or there are entries in gr.crit that don't match any file names.

gr.cols
A character vector, giving one color per group.

fmt
A character string giving the format of the data. Consult import2Dspectra for options.

nF2
Integer giving the number of data points in the F2 (x) dimension.

x.unit
A character string giving the units for the F2 dimension (frequency or wavelength corresponding to the x dimension).

y.unit
A character string giving the units for the F1 dimension (frequency or wavelength corresponding to the y dimension).

z.unit
A character string giving the units of the z-axis (some sort of intensity).

descrip
A character string describing the data set.

fileExt
A character string giving the extension of the files to be processed. regex strings can be used. For instance, the default finds files with either ".csv" or ".CSV" as the extension. Matching is done via a grep process, which is greedy.

out.file
A file name. The completed object of S3 class Spectra2D will be written to this file.

debug
Logical. Set to TRUE for troubleshooting when an error is thrown during import.

...
Arguments to be passed to read.table. You MUST supply values for sep, dec and header consistent with your file structure, unless they are the same as the defaults for read.table.

Value

A object of class Spectra2D. An unnamed object of S3 class Spectra2D is also written to out.file. To read it back into the workspace, use new.name <- loadObject(out.file) (loadObject is package R.utils).

Details

files2Spectra2DObject acts on all files in the current working directory with the specified fileExt so there should be no extra files of that type.

gr.crit and Sample Name Gotchas

The matching of gr.crit against the sample file names (in files2SpectraObject) or column headers/sample names (in codematrix2SpectraObject) is done one at a time, in order, using grep. While powerful, this has the potential to lead to some "gotchas" in certain cases, noted below.

Your file system may allow file/sample names which R will not like, and will cause confusing behavior. File/sample names become variables in ChemoSpec, and R does not like things like "-" (minus sign or hyphen) in file/sample names. A hyphen is converted to a period (".") if found, which is fine for a variable name. However, a period in gr.crit is interpreted from the grep point of view, namely a period matches any single character. At this point, things may behave very differently than one might hope. See make.names for allowed characters in R variables and make sure your file/sample names comply.

The entries in gr.crit must be mutually exclusive. For example, if you have files with names like "Control_1" and "Sample_1" and use gr.crit = c("Control", "Sample") groups will be assigned as you would expect. But, if you have file names like "Control_1_Shade" and "Sample_1_Sun" you can't use gr.crit = c("Control", "Sample", "Sun", "Shade") because each criteria is grepped in order, and the "Sun/Shade" phrases, being last, will form the basis for your groups. Because this is a grep process, you can get around this by using regular expressions in your gr.crit argument to specify the desired groups in a mutually exclusive manner. In this second example, you could use gr.crit = c("Control(.*)Sun", "Control(.*)Shade", "Sample(.*)Sun", "Sample(.*)Shade") to have your groups assigned based upon both phrases in the file names.

To summarize, gr.crit is used as a grep pattern, and the file/sample names are the target. Make sure your file/sample names comply with make.names.

Finally, samples whose names are not matched using gr.crit are still incorporated into the Spectra2D object, but they are not assigned a group. Therefore they don't plot, but they do take up space in a plot! A warning is issued in these cases, since one wouldn't normally want a spectrum to be orphaned this way.

All these problems can generally be identified by running sumSpectra once the data is imported.

Advanced Tricks

While argument fileExt appears to be a file extension (from its name and the description elsewhere), it's actually just a grep pattern that you can apply to any part of the file name if you know how to contruct the proper pattern.